Moves ContentSection component to shared
