+++ src/components/Course/Course.js
@@ -1,9 +1,9 @@
-import React from "react";
-import Tags from "./Tags";
-import StyledLink from "../Link";
-import CourseCard from "../CourseCard";
-import ContentSection from "../ContentSection";
-import "./Course.scss";
+import React from 'react';
+import Tags from './Tags';
+import StyledLink from '../Link';
+import CourseCard from '../CourseCard';
+import ContentSection from '../shared/ContentSection';
+import './Course.scss';
 
 const Course = ({ course }) => {
   const { tags, title, description, lessons } = course.frontmatter;
