+++ src/components/Landing/Landing.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import StyledLink from '../Link';
 import Block from '../shared/Block';
-import ContentSection from '../ContentSection';
+import ContentSection from '../shared/ContentSection';
 import TrackCard from './trackCard';
 import SectionTitle from '../SectionTitle';
 import Text from '../Text';
