+++ src/components/shared/ContentSection/ContentSection.scss
